Update initramfs_add_modules to include lvm module
Testing discovered a scenario in which LVM is in use, but the `lvm`
module is not present in the initramfs image. This causes a failure
on the host during the reboot phase, since LVM will never get loaded.

The value for `initramfs_add_modules` has been updated to include the
`lvm` module explicitly.